# docs/api-by-nest/forum.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "默认模块",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/admin/post-report": {
      "get": {
        "summary": "查询帖子举报列表",
        "deprecated": false,
        "description": "",
        "tags": ["admin"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "list": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "postId": {
                                "type": "string"
                              },
                              "userId": {
                                "type": "string"
                              },
                              "reasons": {
                                "type": "array",
                                "items": {
                                  "type": "string"
                                }
                              }
                            }
                          }
                        },
                        "total": {
                          "type": "integer"
                        }
                      },
                      "required": ["list", "total"]
                    }
                  },
                  "required": ["code", "message", "data"]
                },
                "example": {
                  "code": 0,
                  "message": "success",
                  "data": {
                    "list": [
                      {
                        "postId": "p00161",
                        "userId": "u00000",
                        "reasons": ["不好", "magna do ex"]
                      }
                    ],
                    "total": 2
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/comment-report": {
      "get": {
        "summary": "查询评论举报列表",
        "deprecated": false,
        "description": "",
        "tags": ["admin"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "list": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "postId": {
                                "type": "string"
                              },
                              "commentId": {
                                "type": "integer"
                              },
                              "userId": {
                                "type": "string"
                              },
                              "reasons": {
                                "type": "array",
                                "items": {
                                  "type": "string"
                                }
                              }
                            }
                          }
                        },
                        "total": {
                          "type": "integer"
                        }
                      },
                      "required": ["list", "total"]
                    }
                  },
                  "required": ["code", "message", "data"]
                },
                "example": {
                  "code": 0,
                  "message": "success",
                  "data": {
                    "list": [
                      {
                        "postId": "p00162",
                        "commentId": 167,
                        "userId": "u00001",
                        "reasons": ["ea qui ad"]
                      }
                    ],
                    "total": 1
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/unreview-post": {
      "get": {
        "summary": "查询未审核帖子",
        "deprecated": false,
        "description": "",
        "tags": ["admin"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "list": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "pId": {
                                "type": "string"
                              },
                              "userId": {
                                "type": "string"
                              },
                              "pViewCount": {
                                "type": "string"
                              },
                              "pCollectCount": {
                                "type": "string"
                              },
                              "pShareCount": {
                                "type": "string"
                              },
                              "pCommentCount": {
                                "type": "string"
                              },
                              "pContent": {
                                "type": "string"
                              },
                              "pImages": {
                                "type": "array",
                                "items": {
                                  "type": "string"
                                }
                              },
                              "publishTime": {
                                "type": "string"
                              },
                              "isPublic": {
                                "type": "string"
                              },
                              "status": {
                                "type": "integer"
                              },
                              "username": {
                                "type": "string"
                              },
                              "userAvatar": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "pId",
                              "userId",
                              "pViewCount",
                              "pCollectCount",
                              "pShareCount",
                              "pCommentCount",
                              "pContent",
                              "pImages",
                              "publishTime",
                              "isPublic",
                              "status",
                              "username",
                              "userAvatar"
                            ]
                          }
                        },
                        "total": {
                          "type": "integer"
                        }
                      },
                      "required": ["list", "total"]
                    }
                  },
                  "required": ["code", "message", "data"]
                },
                "example": {
                  "code": 0,
                  "message": "success",
                  "data": {
                    "list": [
                      {
                        "pId": "p00151",
                        "userId": "u00000",
                        "pViewCount": "0",
                        "pCollectCount": "0",
                        "pShareCount": "0",
                        "pCommentCount": "0",
                        "pContent": "333",
                        "pImages": [
                          "/uploads/2026-02-23_19-06-08_b5bd1b05b87f102b.jpg",
                          "/uploads/2026-02-23_19-06-08_9dca83933a54520f.jpg",
                          "/uploads/2026-02-23_19-06-08_b88dffdaa419989c.jpg",
                          "/uploads/2026-02-23_19-06-08_6cf0751314fceea4.jpg",
                          "/uploads/2026-02-23_19-06-08_860e28787374c51e.jpg",
                          "/uploads/2026-02-23_19-06-08_fd3d2bef31539cf8.jpg",
                          "/uploads/2026-02-23_19-06-08_3dab59a59674dfb0.jpg"
                        ],
                        "publishTime": "2026-02-23T11:06:08.000Z",
                        "isPublic": "true",
                        "status": 0,
                        "username": "默认名字1",
                        "userAvatar": "/uploads/2026-02-23_19-06-37_e6ab978fa279b211.jpg"
                      },
                      {
                        "pId": "p00150",
                        "userId": "u00000",
                        "pViewCount": "0",
                        "pCollectCount": "0",
                        "pShareCount": "0",
                        "pCommentCount": "0",
                        "pContent": "222",
                        "pImages": [
                          "/uploads/2026-02-23_18-41-11_f342cb72f9bf33a1.jpg"
                        ],
                        "publishTime": "2026-02-23T10:41:11.000Z",
                        "isPublic": "true",
                        "status": 0,
                        "username": "默认名字1",
                        "userAvatar": "/uploads/2026-02-23_19-06-37_e6ab978fa279b211.jpg"
                      },
                      {
                        "pId": "p00149",
                        "userId": "u00000",
                        "pViewCount": "0",
                        "pCollectCount": "0",
                        "pShareCount": "0",
                        "pCommentCount": "0",
                        "pContent": "111",
                        "pImages": [
                          "/uploads/2026-02-23_18-40-47_7a4f4398e8e2ed5a.jpg"
                        ],
                        "publishTime": "2026-02-23T10:40:47.000Z",
                        "isPublic": "true",
                        "status": 0,
                        "username": "默认名字1",
                        "userAvatar": "/uploads/2026-02-23_19-06-37_e6ab978fa279b211.jpg"
                      },
                      {
                        "pId": "p00148",
                        "userId": "u00000",
                        "pViewCount": "4",
                        "pCollectCount": "1",
                        "pShareCount": "0",
                        "pCommentCount": "0",
                        "pContent": "案发",
                        "pImages": [
                          "/uploads/2026-02-22_19-25-53_d2fbf19d5527d343.jpg",
                          "/uploads/2026-02-22_19-25-57_cb9448de98ba2bd9.jpg",
                          "/uploads/2026-02-22_19-26-00_1abdc69da10109c8.jpg",
                          "/uploads/2026-02-22_19-26-04_cf0257b759944319.jpg",
                          "/uploads/2026-02-22_19-26-08_734c63663a52690d.jpg",
                          "/uploads/2026-02-22_19-26-12_e36c58dd46ab68f8.jpg",
                          "/uploads/2026-02-22_19-26-17_18aeba82e57dad24.jpg"
                        ],
                        "publishTime": "2026-02-22T11:25:50.000Z",
                        "isPublic": "true",
                        "status": 0,
                        "username": "默认名字1",
                        "userAvatar": "/uploads/2026-02-23_19-06-37_e6ab978fa279b211.jpg"
                      },
                      {
                        "pId": "p00147",
                        "userId": "u00000",
                        "pViewCount": "0",
                        "pCollectCount": "0",
                        "pShareCount": "0",
                        "pCommentCount": "0",
                        "pContent": "是高手高高手",
                        "pImages": [
                          "/uploads/2026-02-22_19-24-59_0e2d6e29e940d1d9.jpg",
                          "/uploads/2026-02-22_19-25-00_6558300d5cb3a59d.jpg",
                          "/uploads/2026-02-22_19-25-00_6b0f82f279773817.jpg",
                          "/uploads/2026-02-22_19-25-00_5b7357f71497dbe3.jpg",
                          "/uploads/2026-02-22_19-25-00_15bb8d66fe9f2369.jpg",
                          "/uploads/2026-02-22_19-25-00_6848175ac1ddf3db.jpg",
                          "/uploads/2026-02-22_19-25-00_92697e1c1aafbc09.jpg"
                        ],
                        "publishTime": "2026-02-22T11:24:59.000Z",
                        "isPublic": "true",
                        "status": 0,
                        "username": "默认名字1",
                        "userAvatar": "/uploads/2026-02-23_19-06-37_e6ab978fa279b211.jpg"
                      },
                      {
                        "pId": "p00146",
                        "userId": "u00000",
                        "pViewCount": "0",
                        "pCollectCount": "0",
                        "pShareCount": "0",
                        "pCommentCount": "0",
                        "pContent": "啊发发发的",
                        "pImages": [
                          "/uploads/2026-02-22_19-24-32_6b230b9e9664cb09.jpg",
                          "/uploads/2026-02-22_19-24-32_56726a6729921dc5.jpg",
                          "/uploads/2026-02-22_19-24-32_05c90c168863a803.jpg",
                          "/uploads/2026-02-22_19-24-32_1b56a9b43feb62b7.jpg",
                          "/uploads/2026-02-22_19-24-33_b0aab7d0ddb3a9a4.jpg"
                        ],
                        "publishTime": "2026-02-22T11:24:32.000Z",
                        "isPublic": "true",
                        "status": 0,
                        "username": "默认名字1",
                        "userAvatar": "/uploads/2026-02-23_19-06-37_e6ab978fa279b211.jpg"
                      },
                      {
                        "pId": "p00145",
                        "userId": "u00000",
                        "pViewCount": "2",
                        "pCollectCount": "0",
                        "pShareCount": "0",
                        "pCommentCount": "0",
                        "pContent": "add",
                        "pImages": [
                          "/uploads/2026-02-22_19-10-35_2289a32c20502a4a.jpg",
                          "/uploads/2026-02-22_19-10-35_4ecb04a6396e0252.jpg",
                          "/uploads/2026-02-22_19-10-35_dcefd0b6215eb331.jpg",
                          "/uploads/2026-02-22_19-10-35_db80953e4960d064.jpg",
                          "/uploads/2026-02-22_19-10-35_4dca31e043643658.jpg",
                          "/uploads/2026-02-22_19-10-36_1ac05e3936e2570f.jpg",
                          "/uploads/2026-02-22_19-10-36_997ce9e5f58e8a16.jpg"
                        ],
                        "publishTime": "2026-02-22T11:10:35.000Z",
                        "isPublic": "true",
                        "status": 0,
                        "username": "默认名字1",
                        "userAvatar": "/uploads/2026-02-23_19-06-37_e6ab978fa279b211.jpg"
                      },
                      {
                        "pId": "p00144",
                        "userId": "u00000",
                        "pViewCount": "0",
                        "pCollectCount": "0",
                        "pShareCount": "0",
                        "pCommentCount": "0",
                        "pContent": "654",
                        "pImages": [
                          "/uploads/2026-02-22_19-09-42_3c952642c51b73fe.jpg",
                          "/uploads/2026-02-22_19-09-42_dd7a7445f2eb9c1b.jpg",
                          "/uploads/2026-02-22_19-09-42_ff6200d80f441bff.jpg",
                          "/uploads/2026-02-22_19-09-42_550bc5ca3e6ddab6.jpg"
                        ],
                        "publishTime": "2026-02-22T11:09:42.000Z",
                        "isPublic": "true",
                        "status": 0,
                        "username": "默认名字1",
                        "userAvatar": "/uploads/2026-02-23_19-06-37_e6ab978fa279b211.jpg"
                      },
                      {
                        "pId": "p00143",
                        "userId": "u00000",
                        "pViewCount": "0",
                        "pCollectCount": "0",
                        "pShareCount": "0",
                        "pCommentCount": "0",
                        "pContent": "216",
                        "pImages": [
                          "/uploads/2026-02-22_19-09-31_48b570f56d507e9b.jpg",
                          "/uploads/2026-02-22_19-09-31_0a923254eb6c8a1b.jpg",
                          "/uploads/2026-02-22_19-09-31_1c7ae54431add002.jpg"
                        ],
                        "publishTime": "2026-02-22T11:09:31.000Z",
                        "isPublic": "true",
                        "status": 0,
                        "username": "默认名字1",
                        "userAvatar": "/uploads/2026-02-23_19-06-37_e6ab978fa279b211.jpg"
                      },
                      {
                        "pId": "p00142",
                        "userId": "u00000",
                        "pViewCount": "0",
                        "pCollectCount": "0",
                        "pShareCount": "0",
                        "pCommentCount": "0",
                        "pContent": "4654",
                        "pImages": [
                          "/uploads/2026-02-22_19-09-22_804956bbc535166b.jpg",
                          "/uploads/2026-02-22_19-09-22_d6fd2ba2f41553d9.jpg"
                        ],
                        "publishTime": "2026-02-22T11:09:22.000Z",
                        "isPublic": "true",
                        "status": 0,
                        "username": "默认名字1",
                        "userAvatar": "/uploads/2026-02-23_19-06-37_e6ab978fa279b211.jpg"
                      }
                    ],
                    "total": 114
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/admin/login": {
      "post": {
        "summary": "登录",
        "deprecated": false,
        "description": "",
        "tags": ["admin"],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": ["email", "password"]
              },
              "examples": {}
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "accessToken": {
                          "type": "string"
                        }
                      },
                      "required": ["accessToken"]
                    }
                  },
                  "required": ["code", "message", "data"]
                },
                "example": {
                  "code": 0,
                  "message": "success",
                  "data": {
                    "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6InUwMDAwMCIsImlhdCI6MTc3MzE0ODAzOSwiZXhwIjoxNzczMTQ4OTM5fQ.M2yUH0jQxYI2QvR0sCkwMMI8HlDl192k827eWAAqe0k"
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/admin/post-review": {
      "post": {
        "summary": "修改帖子状态",
        "deprecated": false,
        "description": "",
        "tags": ["admin"],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "postId": {
                    "type": "string"
                  },
                  "status": {
                    "type": "number",
                    "enum": [1, 2]
                  },
                  "reason": {
                    "type": "string"
                  },
                  "punishTime": {
                    "type": "number"
                  }
                },
                "required": ["postId", "status", "reason", "punishTime"]
              },
              "examples": {}
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["code", "message"]
                },
                "example": {
                  "code": 0,
                  "message": "success"
                }
              }
            },
            "headers": {}
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string"
                    },
                    "statusCode": {
                      "type": "integer"
                    }
                  },
                  "required": ["message", "error", "statusCode"]
                },
                "example": {
                  "message": "权限不足",
                  "error": "Forbidden",
                  "statusCode": 403
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [],
  "security": []
}